rename file paths to lowercase
