+++ src/views/docs.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Page from "../components/Page";
+import Page from "../components/page";
 
 const Docs = function () {
   return (
